feat(filter): refactor to remove search filter file
