+++ packages/paste-website/stories/FilterGroup.stories.tsx
@@ -2,7 +2,6 @@ import * as React from "react";
 
 import { CustomDateFilterGroup } from "../src/component-examples/filter-group/components/CustomDateFilterGroup";
 import { DefaultFilterGroup } from "../src/component-examples/filter-group/components/DefaultFilterGroup";
-import { SearchFilterGroup } from "../src/component-examples/filter-group/components/SearchFilterGroup";
 import { STATIC_TABLE_DATA } from "../src/component-examples/filter-group/constants";
 
 export default {
@@ -11,6 +10,6 @@ export default {
 
 export const DefaultFilterGroupExample = (): JSX.Element => <DefaultFilterGroup data={STATIC_TABLE_DATA} />;
 
-export const SearchFilterGroupExample = (): JSX.Element => <SearchFilterGroup data={STATIC_TABLE_DATA} />;
+export const SearchFilterGroupExample = (): JSX.Element => <DefaultFilterGroup data={STATIC_TABLE_DATA} withSearch />;
 
 export const CustomDateFilterGroupExample = (): JSX.Element => <CustomDateFilterGroup data={STATIC_TABLE_DATA} />;
